requirements.txt
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
pandas
numpy
joblib
dvc
Pillow==2.2.0
gunicorn
Keras==2.3.0
tensorflow==2.0.0
#opencv-python-headless
h5py==2.10.0
PyYAML
pandas
